[✨Feat] #49 - 매장 상세 뷰 StickyHeader 구현
